requirements.txt
streamlit
SpeechRecognition
pyttsx3
python-dotenv
google-generativeai
pyaudio
